ZeroShot/docling_page_wise_pdf_converter/requirements.txt
# Core dependencies
docling
docling-core

# Data processing and conversion
pandas
pyyaml
beautifulsoup4

# Image processing
Pillow

# XML processing
lxml # for better XML/HTML handling

# Type hints
typing-extensions

# Select files in explorer view
tkinter